int to float

+++ LD50-Game/Assets/Scripts/Scriptables/FuelControllerData.cs
@@ -6,13 +6,13 @@ namespace Scriptables
     public class FuelControllerData : ScriptableObject
     {
         [SerializeField] private int m_speedUpgradeAfter;
-        [SerializeField] private int m_speedUpgradeAmount;
+        [SerializeField] private float m_speedUpgradeAmount;
         [SerializeField] private int m_tankUpgradeAfter;
-        [SerializeField] private int m_tankUpgradeAmount;
+        [SerializeField] private float m_tankUpgradeAmount;
 
         public int SpeedUpgradeAfter => this.m_speedUpgradeAfter;
-        public int SpeedUpgradeAmount => this.m_speedUpgradeAmount;
+        public float SpeedUpgradeAmount => this.m_speedUpgradeAmount;
         public int TankUpgradeAfter => this.m_tankUpgradeAfter;
-        public int TankUpgradeAmount => this.m_tankUpgradeAmount;
+        public float TankUpgradeAmount => this.m_tankUpgradeAmount;
     }
 }
